Spring_Practice/Practice#5/item-service/src/main/java/hello/itemservice/dmain/item/ItemRepository.java
package hello.itemservice.dmain.item;


import org.springframework.stereotype.Repository;

import javax.servlet.http.PushBuilder;
import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

@Repository
public class ItemRepository {

    private static Map<Long, Item> store = new HashMap<>();
    private static long sequence = 0L;

    public Item save(Item item) {
        item.setId(++sequence);
        store.put(item.getId(), item);
        return item;
    }

    public Item findById(Long id) {
        return store.get(id);
    }

    public List<Item> findAll() {
        return new ArrayList<>(store.values());
    }

    public void update(Long itemId, Item updateParam) {
        Item targetItem = findById(itemId);
        targetItem.setItemName(updateParam.getItemName());
        targetItem.setPrice(updateParam.getPrice());
        targetItem.setQuantity(updateParam.getQuantity());
    }

    public void clearStore() {
        store.clear();
    }
}
